+++ test/unit/fuzz-tools.test.js
@@ -297,6 +297,7 @@ import * as cdiSeverityV317 from '../../lib/cdi-severity-v317.js';
 import * as laEsophagitisV318 from '../../lib/la-esophagitis-v318.js';
 import * as ccsAnginaV319 from '../../lib/ccs-angina-v319.js';
 import * as clavienDindoV320 from '../../lib/clavien-dindo-v320.js';
+import * as hincheyV321 from '../../lib/hinchey-v321.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -539,6 +540,7 @@ const MODULES = {
   'la-esophagitis-v318.js': laEsophagitisV318,
   'ccs-angina-v319.js': ccsAnginaV319,
   'clavien-dindo-v320.js': clavienDindoV320,
+  'hinchey-v321.js': hincheyV321,
   'tb-testing.js': tbTesting,
 };